snn.functional refactored
