+++ snntorch/functional/__init__.py
@@ -1,0 +1,4 @@
+import torch
+from .acc import *
+from .loss import *
+from .reg import *
