+++ snntorch/functional/loss.py
@@ -1,9 +1,12 @@
 import torch
 import torch.nn as nn
 from snntorch import spikegen
-import numpy as np
 
-#### Note: when adding new loss or regularization functions, be sure to update criterion_dict / reg_dict in backprop.py
+###############################################################
+# When adding new loss/reg functions, update criterion_dict   #
+#                and reg_dict in backprop.py                  #
+###############################################################
+
 
 dtype = torch.float
 
@@ -125,7 +128,7 @@ class ce_count_loss(LossFunctions):
 
         if self.population_code:
             _, _, num_outputs = self._prediction_check(spk_out)
-            spike_count = _population_code(spk_out, self.num_classes, num_outputs)
+            spike_count = self._population_code(spk_out, self.num_classes, num_outputs)
         else:
             spike_count = torch.sum(spk_out, 0)  # B x C
         log_p_y = log_softmax_fn(spike_count)
@@ -236,7 +239,7 @@ class mse_count_loss(LossFunctions):
                 on_target=on_target,
                 off_target=off_target,
             )
-            spike_count = _population_code(spk_out, self.num_classes, num_outputs)
+            spike_count = self._population_code(spk_out, self.num_classes, num_outputs)
 
         loss = loss_fn(spike_count, spike_count_target)
         return loss / num_steps
@@ -328,83 +331,3 @@ class mse_membrane_loss(LossFunctions):
 #     loss = loss_fn(log_p_y, targets)
 
 #     return loss
-
-
-def accuracy_rate(spk_out, targets, population_code=False, num_classes=False):
-    """Use spike count to measure accuracy.
-
-    :param spk_out: Output spikes of shape [num_steps x batch_size x num_outputs]
-    :type spk_out: torch.Tensor
-
-    :param targets: Target tensor (without one-hot-encoding) of shape [batch_size]
-    :type targets: torch.Tensor
-
-    :return: accuracy
-    :rtype: numpy.float64
-    """
-    if population_code:
-        _, _, num_outputs = _prediction_check(spk_out)
-        _, idx = _population_code(spk_out, num_classes, num_outputs).max(1)
-
-    else:
-        _, idx = spk_out.sum(dim=0).max(1)
-    accuracy = np.mean((targets == idx).detach().cpu().numpy())
-
-    return accuracy
-
-
-class l1_rate_sparsity:
-    """L1 regularization using total spike count as the penalty term.
-    Lambda is a scalar factor for regularization."""
-
-    def __init__(self, Lambda=1e-5):
-        self.Lambda = Lambda
-        self.__name__ = "l1_rate_sparsity"
-
-    def __call__(self, spk_out):
-        return self.Lambda * torch.sum(spk_out)
-
-
-# # def l2_sparsity(mem_out, Lambda=1e-6):
-# #     """L2 regularization using accumulated membrane potential as the penalty term."""
-# #     return Lambda * (torch.sum(mem_out)**2)
-
-
-def _prediction_check(spk_out):
-    device = "cpu"
-    if spk_out.is_cuda:
-        device = "cuda"
-
-    num_steps = spk_out.size(0)
-    num_outputs = spk_out.size(-1)
-
-    return device, num_steps, num_outputs
-
-
-def _population_code(spk_out, num_classes, num_outputs):
-    """Count up spikes sequentially from output classes."""
-    if not num_classes:
-        raise Exception(
-            "``num_classes`` must be specified if ``population_code=True``."
-        )
-    if num_outputs % num_classes:
-        raise Exception(
-            f"``num_outputs {num_outputs} must be a factor of num_classes {num_classes}."
-        )
-    device = "cpu"
-    if spk_out.is_cuda:
-        device = "cuda"
-    pop_code = torch.zeros(tuple([spk_out.size(1)] + [num_classes])).to(device)
-    for idx in range(num_classes):
-        pop_code[:, idx] = (
-            spk_out[
-                :,
-                :,
-                int(num_outputs * idx / num_classes) : int(
-                    num_outputs * (idx + 1) / num_classes
-                ),
-            ]
-            .sum(-1)
-            .sum(0)
-        )
-    return pop_code
